// src/types/types.ts
export interface MovieResult {
  adult: boolean;
  backdrop_path: string | null;
  genre_ids: number[];
  id: number;
  original_language: string;
  original_title: string;
  overview: string;
  popularity: number;
  poster_path: string | null;
  release_date: string;
  title: string;
  video: boolean;
  vote_average: number;
  vote_count: number;
}

export interface AIAnswerResponse {
  query: string;
  answer: string;
  images: string[];
  results: SearchResult[];
  auto_parameters: AutoParameters;
  response_time: string;
}

export interface SearchResult {
  title: string;
  url: string;
  content: string;
  score: number;
  raw_content: string | null;
  favicon: string;
}

export interface AutoParameters {
  topic: string;
  search_depth: string;
}